backend/quizapp/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
asttokens==3.0.1
certifi==2025.11.12
decorator==5.2.1
distro==1.9.0
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
exceptiongroup==1.3.1
executing==2.2.1
groq==0.37.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
sniffio==1.3.1
sqlparse==0.5.4
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
typing-inspection==0.4.2
wcwidth==0.2.14
